[Refactored] Fuse code into its own class
