+++ Packages/Core/Editor/Utils/AtomFuser.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 765dd6b85b7be84428e3e5726361eac8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
